backend/requirements.txt
djangorestframework==3.11.2
Django==4.1.1
Pygments==2.7.4
Markdown==3.2.1
coreapi==2.3.3
psycopg2-binary==2.9.3
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
PyYAML==5.4
pytz==2022.2.1 